requirement.txt
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.1.1
cloudpickle==2.2.0
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.6.1
fasteners==0.18
google-api-core==2.10.1
google-apitools==0.5.31
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-aiplatform==1.17.1
google-cloud-appengine-logging==1.1.4
google-cloud-audit-log==0.2.4
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.13.2
google-cloud-bigtable==1.7.2
google-cloud-core==2.3.2
google-cloud-datastore==1.15.5
google-cloud-dlp==3.9.0
google-cloud-language==1.3.2
google-cloud-logging==3.2.2
google-cloud-pubsub==2.13.7
google-cloud-pubsublite==1.5.0
google-cloud-recommendations-ai==0.7.1
google-cloud-resource-manager==1.6.1
google-cloud-spanner==1.19.3
google-cloud-storage==2.5.0
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.49.1
grpcio-gcp==0.2.2
grpcio-status==1.48.2
hdfs==2.7.0
httplib2==0.20.4
idna==3.4
numpy==1.21.6
oauth2client==4.1.3
orjson==3.8.0
overrides==6.2.0
packaging==21.3
pandas==1.3.5
proto-plus==1.22.1
protobuf==3.20.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
whatever==0.6